Edge inference example refactored as the other ones
