+++ C/examples/edgeinference/edgeinference.cpp
@@ -38,140 +38,17 @@
 #include <ff/dff.hpp>
 #include <torch/torch.h>
 
-#include "process-frame.hpp"
+#include "C/utils/net.hpp"
 #include "C/utils/serialize.hpp"
+#include "C/dfl/video_inference.hpp"
 
 using namespace ff;
-
-struct edgeMsg_t {
-    edgeMsg_t() {}
-
-    edgeMsg_t(edgeMsg_t *t) {
-        str = std::string(t->str);
-        frame_n = t->frame_n;
-        //nodeTime = t->nodeTime;
-        nodeFrameRate = t->nodeFrameRate;
-        bboxes = t->bboxes;
-    }
-
-    std::string str;
-    unsigned long frame_n;
-    // std::chrono::steady_clock::time_point nodeTime;
-    float nodeFrameRate;
-    std::vector <BboxInfo> bboxes;
-
-    template<class Archive>
-    void serialize(Archive &archive) {
-        archive(str, frame_n, nodeFrameRate, bboxes);
-    }
-
-};
-
-struct EdgeNode : ff_monode_t<edgeMsg_t> {
-    EdgeNode(std::string workerName, torch::jit::script::Module model, std::string video_fname) : workerName(
-            workerName), model(model), video_fname(video_fname) {}
-
-    int svc_init() {
-        std::cout << "[" << workerName << "] Starting to process video..." << video_fname << std::endl;
-        cap = cv::VideoCapture(video_fname);
-        start = std::chrono::steady_clock::now();
-        frame_count = 0;
-        return 0;
-    }
-
-    void svc_end() {
-        // 	// Close the file?
-        cap.release();
-        std::cout << "Video closed, process finished. " << std::endl;
-    }
-
-    edgeMsg_t *svc(edgeMsg_t *) {
-
-        while (cap.isOpened()) {
-            frame_count++;
-            std::cout << "[" << workerName << "] Processing frame..." << std::endl;
-            cap.read(frame);
-            if (frame.empty()) {
-                std::cout << "[" << workerName << "] Read frame failed!" << std::endl;
-                break;
-            }
-
-            torch::Tensor imgTensor = imgToTensor(frame);
-            torch::Tensor preds = model.forward({imgTensor}).toTuple()->elements()[0].toTensor();
-
-            torch::Tensor detections = non_max_suppression(preds, 0.5, 0.3);
-            std::vector <BboxInfo> bboxes = compute_bboxes(frame, detections);
-
-            //process_results(bboxes);
-            end = std::chrono::steady_clock::now();
-            elapsed = end - start;
-            uint64_t elapsed_sec = std::chrono::duration_cast<std::chrono::seconds>(elapsed).count();
-
-            edgeMsg_t *task = new edgeMsg_t;
-            task->bboxes = bboxes;
-            task->frame_n = frame_count;
-            //task->nodeTime=end;
-            task->str = "Msg";
-            ff_send_out(task);
-        }
-        return EOS;
-    }
-
-    const std::string workerName, video_fname;
-    torch::jit::script::Module model;
-    std::chrono::steady_clock::time_point start, end;
-    std::chrono::steady_clock::duration elapsed;
-    unsigned long frame_count;
-    cv::VideoCapture cap;
-    cv::Mat frame;
-
-};
-
-struct level1Gatherer : ff_minode_t<edgeMsg_t> {
-    edgeMsg_t *svc(edgeMsg_t *t) {
-        //t->str += std::string(" World");
-        auto results = t->bboxes;
-        for (auto bbox: results) {
-            man_down(bbox);
-            if (isMandown == true) {
-                //std::cout << bbox.p1 << " " << bbox.p2 << " mandown"<<std::endl;
-                edgeMsg_t *task = new edgeMsg_t;
-                task->bboxes.push_back(bbox);
-                task->str = "Msga";
-                ff_send_out(task);
-                //}
-
-            }
-
-        }
-        delete t;
-        return GO_ON;
-    }
-};
-
-
-struct HelperNode : ff_monode_t<edgeMsg_t> {
-    edgeMsg_t *svc(edgeMsg_t *task) { return task; }
-};
-
-
-struct level0Gatherer : ff_minode_t<edgeMsg_t> {
-    edgeMsg_t *svc(edgeMsg_t *task) {
-        auto results = task->bboxes;
-        for (auto bbox: results) {
-            //std::cout << "mandown in root "<< bbox.p1 << " " << bbox.p2 << std::endl;
-        }
-        // std::cerr << "level0Gatherer: from (" << get_channel_id() << ") " << t->str << "frame count " << t->S.f;
-        return GO_ON;
-    }
-};
-
 
 int main(int argc, char *argv[]) {
     timer chrono = timer("Total execution time");
 
     std::string groupName = "W0";
-    std::string loggerName = "W0";
+    const std::string loggerName = "W0";
 
 #ifndef DISABLE_FF_DISTRIBUTED
     for (int i = 0; i < argc; i++)
@@ -186,77 +63,75 @@ int main(int argc, char *argv[]) {
     }
 #endif
 
-    size_t nL = 3;
-    size_t pL = 1; // 1 client per group
-    int forcecpu = 0;
-    std::string inmodel = "../../../data/yolov5n.torchscript";
-    std::string infile = "../../../data/Ranger_Roll_m.mp4";
+    int num_workers{3};     // Number of workers
+    int num_groups{1};      // Number of workers per group
+    int forcecpu{0};        // Force the execution on the CPU
+    char *data_path;        // Patch to the dataset (absolute or with respect to build directory)
+    char *inmodel;          // Path to a TorchScript representation of a DNN
 
     if (argc >= 2) {
         if (strcmp(argv[1], "-h") == 0) {
             if (groupName.compare(loggerName) == 0)
                 std::cout
-                        << "Usage: edgeinference [forcecpu=0/1] [groups=3] [clients/group=1] [model_path] [data_path]\n";
+                        << "Usage: edgeinference [forcecpu=0/1] [data_path] [groups=3] [clients/group=1] [model_path]\n";
             exit(0);
         } else
             forcecpu = atoi(argv[1]);
     }
     if (argc >= 3)
-        nL = atoi(argv[2]);
+        data_path = argv[2];
     if (argc >= 4)
-        pL = atoi(argv[3]);
+        num_workers = atoi(argv[2]);
     if (argc >= 5)
-        inmodel = argv[4];
+        num_groups = atoi(argv[4]);
     if (argc >= 6)
-        infile = argv[5];
+        inmodel = argv[5];
+    if (groupName.compare(loggerName) == 0)
+        std::cout << "Infering on " << num_workers << " cameras." << std::endl;
 
     torch::DeviceType device_type;
     if (torch::cuda::is_available() && !forcecpu) {
-        //std::cout << "CUDA available! Training on GPU." << std::endl;
+        if (groupName.compare(loggerName) == 0)
+            std::cout << "CUDA available! Training on GPU." << std::endl;
         device_type = torch::kCUDA;
     } else {
-        //std::cout << "Training on CPU." << std::endl;
+        if (groupName.compare(loggerName) == 0)
+            std::cout << "Training on CPU." << std::endl;
         device_type = torch::kCPU;
     }
     torch::Device device(device_type);
-
     torch::cuda::manual_seed_all(42);
 
-    torch::jit::script::Module model;
-    try {
-        model = torch::jit::load(inmodel);
-    } catch (const c10::Error &e) {
-        std::cerr << "Error loading the model\n" << std::endl
-                  << "Details:" << std::endl
-                  << e.what() << std::endl;
+    if (groupName.compare(loggerName) == 0)
+        std::cout << "Checking data existence..." << std::endl;
+    if (!std::filesystem::exists(data_path)) {
+        error("Video file cannot be found at path: %s\n", data_path);
         return -1;
     }
-    //  TBD one or more file per node
-    //if (argc>3) infile = argv[2];
-    if (std::filesystem::exists(infile) == false) {
-        std::cerr << "Video file cannot be found at path: " << argv[2] << std::endl;
-        return -1;
-    } //else infile=std::string(argv[2]);
 
+    if (groupName.compare(loggerName) == 0)
+        std::cout << "Cameras creation..." << std::endl;
     ff_a2a a2a;
-    level0Gatherer root;
+    level0Gatherer <edgeMsg_t> root;
     std::vector < ff_node * > globalLeft;
-
-    for (size_t i = 0; i < nL; ++i) {
+    for (int i = 0; i < num_workers; ++i) {
         ff_pipeline *pipe = new ff_pipeline;   // <---- To be removed and automatically added
         ff_a2a *local_a2a = new ff_a2a;
         pipe->add_stage(local_a2a, true);
         std::vector < ff_node * > localLeft;
-        for (size_t j = 0; j < pL; ++j)
-            localLeft.push_back(new EdgeNode("W(" + std::to_string(i) + "," + std::to_string(j) + ")", model, infile));
+        for (int j = 0; j < num_groups; ++j) {
+            Net <torch::jit::Module> *model = new Net<torch::jit::Module>(inmodel);
+            localLeft.push_back(
+                    new EdgeNode < Net < torch::jit::Module > *, edgeMsg_t > (
+                            "W(" + std::to_string(i) + "," + std::to_string(j) + ")", model, data_path));
+        }
         local_a2a->add_firstset(localLeft, 0, true);
-        local_a2a->add_secondset<ff_comb>({new ff_comb(new level1Gatherer, new HelperNode, true, true)});
+        local_a2a->add_secondset<ff_comb>(
+                {new ff_comb(new level1Gatherer <edgeMsg_t>, new HelperNode <edgeMsg_t>, true, true)});
         globalLeft.push_back(pipe);
-        // create here Gi groups for each a2a(i) i>0
         auto g = a2a.createGroup("W" + std::to_string(i + 1));
         g << pipe;
     }
-
     a2a.add_firstset(globalLeft, true);
     a2a.add_secondset<ff_node>({&root});
     a2a.createGroup("W0") << root;
